requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
aeneas==1.7.3.0
    # via audiobooksyncer (pyproject.toml)
beautifulsoup4==4.12.3
    # via aeneas
bertalign @ git+https://github.com/atlantis-11/bertalign.git
    # via audiobooksyncer (pyproject.toml)
certifi==2024.7.4
    # via
    #   httpx
    #   requests
chardet==3.0.4
    # via httpx
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via nltk
faiss-gpu==1.7.2
    # via bertalign
ffmpeg-python==0.2.0
    # via audiobooksyncer (pyproject.toml)
filelock==3.15.4
    # via
    #   huggingface-hub
    #   torch
    #   transformers
    #   triton
fsspec==2024.6.1
    # via
    #   huggingface-hub
    #   torch
future==1.0.0
    # via ffmpeg-python
googletrans==4.0.0rc1
    # via bertalign
h11==0.9.0
    # via httpcore
h2==3.2.0
    # via httpcore
hpack==3.0.0
    # via h2
hstspreload==2024.8.1
    # via httpx
httpcore==0.9.1
    # via httpx
httpx==0.13.3
    # via googletrans
huggingface-hub==0.24.5
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
hyperframe==5.2.0
    # via h2
idna==2.10
    # via
    #   httpx
    #   requests
jinja2==3.1.4
    # via torch
joblib==1.4.2
    # via
    #   nltk
    #   scikit-learn
llvmlite==0.39.1
    # via numba
lxml==5.3.0
    # via aeneas
markupsafe==2.1.5
    # via jinja2
more-itertools==10.4.0
    # via openai-whisper
mpmath==1.3.0
    # via sympy
networkx==3.2.1
    # via torch
nltk==3.9.1
    # via sentence-transformers
numba==0.56.4
    # via
    #   bertalign
    #   openai-whisper
numpy==1.23.5
    # via
    #   aeneas
    #   numba
    #   openai-whisper
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   torchvision
    #   transformers
nvidia-cublas-cu12==12.1.3.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.1.105
    # via torch
nvidia-cuda-nvrtc-cu12==12.1.105
    # via torch
nvidia-cuda-runtime-cu12==12.1.105
    # via torch
nvidia-cudnn-cu12==8.9.2.26
    # via torch
nvidia-cufft-cu12==11.0.2.54
    # via torch
nvidia-curand-cu12==10.3.2.106
    # via torch
nvidia-cusolver-cu12==11.4.5.107
    # via torch
nvidia-cusparse-cu12==12.1.0.106
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.20.5
    # via torch
nvidia-nvjitlink-cu12==12.6.20
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
nvidia-nvtx-cu12==12.1.105
    # via torch
openai-whisper @ git+https://github.com/openai/whisper.git@v20231117
    # via audiobooksyncer (pyproject.toml)
packaging==24.1
    # via
    #   huggingface-hub
    #   transformers
pillow==10.4.0
    # via torchvision
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   transformers
rapidfuzz==3.9.6
    # via thefuzz
regex==2024.7.24
    # via
    #   nltk
    #   sentence-splitter
    #   tiktoken
    #   transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   tiktoken
    #   transformers
rfc3986==1.5.0
    # via httpx
safetensors==0.4.4
    # via transformers
scikit-learn==1.5.1
    # via sentence-transformers
scipy==1.13.1
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-splitter==1.4
    # via bertalign
sentence-transformers==2.2.2
    # via bertalign
sentencepiece==0.2.0
    # via sentence-transformers
sniffio==1.3.1
    # via
    #   httpcore
    #   httpx
soupsieve==2.6
    # via beautifulsoup4
sympy==1.13.2
    # via torch
thefuzz==0.22.1
    # via audiobooksyncer (pyproject.toml)
threadpoolctl==3.5.0
    # via scikit-learn
tiktoken==0.7.0
    # via openai-whisper
tokenizers==0.19.1
    # via transformers
torch==2.3.1
    # via
    #   openai-whisper
    #   sentence-transformers
    #   torchvision
torchvision==0.18.1
    # via sentence-transformers
tqdm==4.66.5
    # via
    #   huggingface-hub
    #   nltk
    #   openai-whisper
    #   sentence-transformers
    #   transformers
transformers==4.44.0
    # via sentence-transformers
triton==2.3.1
    # via
    #   openai-whisper
    #   torch
typing-extensions==4.12.2
    # via
    #   huggingface-hub
    #   torch
urllib3==2.2.2
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
